STS: Mark avatar support as recommended, but not mandatory, v2
Summary:

According to @tux3 and @dvor's suggestion, support for avatars should be
recommended, but not mandatory for Tox clients.

As @GrayHatter requested, there is defined behaviour for clients that do not
support avatars.

Arguments for:

* Avatars should be an optional feature, since clients can be visually
  appealing without them
* There should be defined behaviour in case where client does not support
  avatars

Arguments against:

* GUI clients without support for avatars might be a bad surprise for people

Status quo:

Currently it is not specified whether clients have to have avatar support.